filter used and getTable functionality moved to another file so that we dont have to make changes in multiple files
